+++ compatibility/src/test/java/io/cucumber/compatibility/docstrings/DocStrings.java
@@ -1,0 +1,12 @@
+package io.cucumber.compatibility.docstrings;
+
+import io.cucumber.java.en.Given;
+
+public class DocStrings {
+
+    @Given("a doc string:")
+    public void docString(String docString) {
+
+    }
+
+}