+++ compatibility/src/test/java/io/cucumber/compatibility/multiplefeaturesreversed/MultipleFeatures.java
@@ -1,0 +1,12 @@
+package io.cucumber.compatibility.multiplefeaturesreversed;
+
+import io.cucumber.java.en.Given;
+
+public class MultipleFeatures {
+
+    @Given("an order for {string}")
+    public void an_order_for_item(String item) {
+
+    }
+
+}