+++ cucumber-core/src/main/java/io/cucumber/core/runner/TestCase.java
@@ -133,6 +133,11 @@ final class TestCase implements io.cucumber.plugin.event.TestCase {
     }
 
     @Override
+    public String getLanguage() {
+        return pickle.getLanguage();
+    }
+
+    @Override
     public List<String> getTags() {
         return pickle.getTags();
     }