+++ cucumber-core/src/main/java/io/cucumber/core/runner/TestCaseState.java
@@ -149,6 +149,11 @@ class TestCaseState implements io.cucumber.core.backend.TestCaseState {
     }
 
     @Override
+    public String getLanguage() {
+        return testCase.getLanguage();
+    }
+
+    @Override
     public Integer getLine() {
         return testCase.getLocation().getLine();
     }