+++ cucumber-java/src/main/java/io/cucumber/java/Scenario.java
@@ -142,6 +142,15 @@ public final class Scenario {
     }
 
     /**
+     * Return the language of the Scenario.
+     * 
+     * @return the language of the Scenario.
+     */
+    public String getLanguage() {
+        return delegate.getLanguage();
+    }
+
+    /**
      * Returns the line in the feature file of the Scenario.
      * <p>
      * If this is a Scenario from Scenario Outlines this will return the line of