+++ cucumber-junit-platform-engine/src/test/java/io/cucumber/junit/platform/engine/TestCaseResultObserverTest.java
@@ -69,6 +69,11 @@ class TestCaseResultObserverTest {
         }
 
         @Override
+        public String getLanguage() {
+            return "en";
+        }
+
+        @Override
         public List<String> getTags() {
             return emptyList();
         }